Fix `{{.evidence}}` references not passing validation
