+++ ghostwriter/modules/reportwriter/forms.py
@@ -4,6 +4,7 @@ from django import forms
 import jinja2
 
 from ghostwriter.modules.reportwriter import prepare_jinja2_env
+from ghostwriter.modules.reportwriter.base import rich_text_template
 
 
 class JinjaRichTextField(forms.CharField):
@@ -19,7 +20,7 @@ class JinjaRichTextField(forms.CharField):
         super().validate(value)
         env, _ = prepare_jinja2_env(debug=True)
         try:
-            env.from_string(value)
+            rich_text_template(env, value)
         except jinja2.TemplateSyntaxError as e:
             line = value.splitlines()[e.lineno - 1]
             raise forms.ValidationError(f"{e} at `{line}`") from e
